(Early) API Version 5
Building these extensions and/or using the new API changes require the ModBot binary that has been added, do not use it in your channel, this is an early, unstable build, database data could be corrupted with use.

These API changes are not final and may change again before the final release.
The extensions have been updated to work only, more changes will be made in the future.

Please report issues encountered with the API changes, most of the changes haven't been tested yet and feedback is appreciated.

To view all of the functions available from the API, add ModBot.exe as a reference and view it in the object browser.